UI: Move mapNode creation logic into the MapPageWrapper.tsx
